Move IconFromUserType and Result to seperate file
